crux-mir: refactor: deduplicate definitions of readMaybeType
